refactor: introduce YamlProcessUnit types and rename to YamlProcessPipeline
