+++ src/libecalc/presentation/yaml/mappers/process_simulation_mapper.py
@@ -37,12 +37,11 @@ from libecalc.presentation.yaml.mappers.fluid_mapper import (
 from libecalc.presentation.yaml.mappers.model import InvalidChartResourceException
 from libecalc.presentation.yaml.yaml_types.components.yaml_expression_type import YamlExpressionType
 from libecalc.presentation.yaml.yaml_types.components.yaml_process_system import (
-    YamlCompressor,
-    YamlCompressorModelChart,
     YamlCompressorStageProcessSystem,
     YamlProcessSimulation,
     YamlSerialProcessSystem,
 )
+from libecalc.presentation.yaml.yaml_types.components.yaml_process_units import YamlCompressor, YamlCompressorModelChart
 from libecalc.presentation.yaml.yaml_types.models import YamlFluidModel
 from libecalc.presentation.yaml.yaml_types.models.yaml_compressor_stages import YamlControlMarginUnits
 from libecalc.presentation.yaml.yaml_types.models.yaml_fluid import YamlCompositionFluidModel, YamlPredefinedFluidModel
