+++ src/libecalc/presentation/yaml/yaml_types/components/yaml_process_system.py
@@ -4,50 +4,13 @@ from pydantic import Field
 
 from libecalc.presentation.yaml.yaml_types import YamlBase
 from libecalc.presentation.yaml.yaml_types.components.yaml_expression_type import YamlExpressionType
-from libecalc.presentation.yaml.yaml_types.models.yaml_compressor_chart import UnitsField, YamlCurve, YamlUnits
-from libecalc.presentation.yaml.yaml_types.models.yaml_compressor_stages import YamlControlMarginUnits
+from libecalc.presentation.yaml.yaml_types.components.yaml_process_units import (
+    ProcessUnitReference,
+    YamlCompressor,
+)
 from libecalc.presentation.yaml.yaml_types.streams.yaml_inlet_stream import YamlInletStream
-from libecalc.presentation.yaml.yaml_types.yaml_data_or_file import DataOrFile
 
 StreamRef = str
-
-
-class YamlControlMargin(YamlBase):
-    unit: YamlControlMarginUnits
-    value: float
-
-
-class YamlCompressorChart(YamlBase):
-    curves: Annotated[
-        DataOrFile[list[YamlCurve]],
-        Field(description="Compressor chart curves, one per speed.", title="CURVES"),
-    ]
-    units: YamlUnits = UnitsField()
-
-
-class YamlCompressorModelChart(YamlBase):
-    type: Literal["COMPRESSOR_CHART"]
-    chart: YamlCompressorChart
-    control_margin: YamlControlMargin
-
-
-ProcessUnitReference = str
-
-
-class YamlCompressor(YamlBase):
-    """
-    A Compressor process unit
-    """
-
-    type: Literal["COMPRESSOR"]
-    name: Annotated[
-        ProcessUnitReference,
-        Field(
-            description="Name of the model. See documentation for more information.",
-            title="NAME",
-        ),
-    ]
-    compressor_model: YamlCompressorModelChart
 
 
 type ProcessSystemReference = str  # TODO: validate correct reference
@@ -151,11 +114,6 @@ class YamlProcessSimulation(YamlBase):
         ),
     ]
 
-
-YamlProcessUnit = Annotated[
-    YamlCompressor,
-    Field(discriminator="type"),
-]
 
 YamlProcessSystem = Annotated[
     YamlSerialProcessSystem | YamlCompressorStageProcessSystem,
